Add SensorDataRepository Interface
* Create interface `SensorDataRepository` with two needed repository methods
* Change existing interface to `SensorDataRepositoryImpl` and implement new interface
* Services use repository interface which's bean is represented by actual implementation
